Add PY_LONG_LONG support from patch contributed by Dmitry Kostjuchenko.

+++ CXX/Extensions.hxx
@@ -426,11 +426,11 @@ namespace Py
         // Sequence methods
         virtual int sequence_length();
         virtual Object sequence_concat( const Object & );
-        virtual Object sequence_repeat( int );
-        virtual Object sequence_item( int );
-        virtual Object sequence_slice( int, int );
-        virtual int sequence_ass_item( int, const Object & );
-        virtual int sequence_ass_slice( int, int, const Object & );
+        virtual Object sequence_repeat( Py_ssize_t );
+        virtual Object sequence_item( Py_ssize_t );
+        virtual Object sequence_slice( Py_ssize_t, Py_ssize_t );
+        virtual int sequence_ass_item( Py_ssize_t, const Object & );
+        virtual int sequence_ass_slice( Py_ssize_t, Py_ssize_t, const Object & );
         
         // Mapping
         virtual int mapping_length();
